Update point schema
We decided to give more weight to communication and report, so there is now a additional section on communication and points in terms of written documents. The final presentation during the interview is more important though still.

I also wouldn't be to strict on the runnable thing. It shouldn't be a binary full or no points. Instead we should ask participants to include a log of their run, then run again on our local machines to see if still works well.